Loggers terminados en todos los service. Validacion del token y RequestHeader terminado y funcionando en todos los controller, YAYY

+++ proyectoFinal/src/main/java/com/dh/dentalClinic/controller/AddressController.java
@@ -2,6 +2,7 @@ package com.dh.dentalClinic.controller;
 import com.dh.dentalClinic.exceptions.BadRequestException;
 import com.dh.dentalClinic.persistence.entities.Address;
 import com.dh.dentalClinic.service.AddressService;
+import com.dh.dentalClinic.user.utils.JWTUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
@@ -10,33 +11,44 @@ import java.util.List;
 @RequestMapping("/addresses")
 public class AddressController {
 
+    @Autowired
+    private JWTUtil jwtUtil;
 
     @Autowired
     AddressService service;
 
     @PostMapping
-    public Address create(@RequestBody Address a) throws BadRequestException {
+    public Address create(@RequestHeader(value = "Authorization") String token, @RequestBody Address a) throws BadRequestException {
+        if (!validarToken(token)) { return null; }
         return service.save(a);
     }
 
     @GetMapping
-    public List<Address> getAll() throws BadRequestException{
+    public List<Address> getAll(@RequestHeader(value = "Authorization") String token) throws BadRequestException{
+        if (!validarToken(token)) { return null; }
         return service.getAll();
     }
 
     @GetMapping("/{id}")
-    public Address getById(@PathVariable Long id) throws BadRequestException{
+    public Address getById(@RequestHeader(value = "Authorization") String token, @PathVariable Long id) throws BadRequestException{
+        if (!validarToken(token)) { return null; }
         return service.getById(id);
     }
 
     @DeleteMapping(value = "{id}")
-    public String delete(@PathVariable Long id) throws BadRequestException{
+    public String delete(@RequestHeader(value = "Authorization") String token, @PathVariable Long id) throws BadRequestException{
+        if (!validarToken(token)) { return null; }
         return service.delete(id);
     }
 
     @PutMapping
-    public String update(@RequestBody Address a) throws BadRequestException{
+    public String update(@RequestHeader(value = "Authorization") String token, @RequestBody Address a) throws BadRequestException{
+        if (!validarToken(token)) { return null; }
         return service.updateAppointment(a);
     }
 
+    private boolean validarToken(String token) {
+        String usuarioId = jwtUtil.getKey(token);
+        return usuarioId != null;
+    }
 }
